requirements.txt
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
streamlit==1.27.0
joblib==1.3.2
google-generativeai==0.3.1
python-dotenv==1.0.0 